html_root_url: move check_html_root_url to new module
The function is re-exported under the old name, so this is transparent
to users of the crate.
