+++ src/lib.rs
@@ -65,11 +65,8 @@ use std::io::Read;
 use std::result;
 
 use itertools::join;
-use semver_parser::range::parse as parse_request;
 use semver_parser::range::{Op, VersionReq};
-use semver_parser::version::parse as parse_version;
 use semver_parser::version::Version;
-use url::Url;
 
 /// The common result type, our errors will be simple strings.
 type Result<T> = result::Result<T, String>;
@@ -186,144 +183,8 @@ macro_rules! assert_markdown_deps_updated {
     };
 }
 
-fn url_matches(value: &str, pkg_name: &str, version: &Version) -> Result<()> {
-    let url = Url::parse(value).map_err(|err| format!("parse error: {}", err))?;
-
-    // We can only reason about docs.rs.
-    if url.domain().is_some() && url.domain() != Some("docs.rs") {
-        return Ok(());
-    }
-
-    // Since docs.rs redirects HTTP traffic to HTTPS, we will ensure
-    // that the scheme is "https" here.
-    if url.scheme() != "https" {
-        return Err(format!("expected \"https\", found {:?}", url.scheme()));
-    }
-
-    let mut path_segments = url
-        .path_segments()
-        .ok_or_else(|| String::from("no path in URL"))?;
-
-    // The package name should not be empty.
-    let name = path_segments
-        .next()
-        .and_then(|path| if path.is_empty() { None } else { Some(path) })
-        .ok_or_else(|| String::from("missing package name"))?;
-
-    // The version number should not be empty.
-    let request = path_segments
-        .next()
-        .and_then(|path| if path.is_empty() { None } else { Some(path) })
-        .ok_or_else(|| String::from("missing version number"))?;
-
-    // Finally, we check that the package name and version matches.
-    if name != pkg_name {
-        Err(format!(
-            "expected package \"{}\", found \"{}\"",
-            pkg_name, name
-        ))
-    } else {
-        // The Rust API Guidelines[1] suggest using an exact version
-        // number, but we have relaxed this a little and allow the
-        // user to specify the version as just "1" or "1.2". We might
-        // make this more strict in the future.
-        //
-        // [1]: https://rust-lang-nursery.github.io/api-guidelines/documentation.html
-        // #crate-sets-html_root_url-attribute-c-html-root
-        parse_request(request)
-            .map_err(|err| format!("could not parse version in URL: {}", err))
-            .and_then(|request| version_matches_request(version, &request))
-    }
-}
-
-/// Check version numbers in `html_root_url` attributes.
-///
-/// This function parses the Rust source file in `path` and looks for
-/// `html_root_url` attributes. Such an attribute must specify a valid
-/// URL and if the URL points to docs.rs, it must be point to the
-/// documentation for `pkg_name` and `pkg_version`.
-///
-/// # Errors
-///
-/// If any attribute fails the check, an `Err` is returned with a
-/// succinct error message. Status information has then already been
-/// printed on `stdout`.
-pub fn check_html_root_url(path: &str, pkg_name: &str, pkg_version: &str) -> Result<()> {
-    let code = read_file(path).map_err(|err| format!("could not read {}: {}", path, err))?;
-    let version = parse_version(pkg_version)
-        .map_err(|err| format!("bad package version {:?}: {}", pkg_version, err))?;
-    let krate: syn::File = syn::parse_str(&code)
-        .map_err(|_| format!("could not parse {}: please run \"cargo build\"", path))?;
-
-    println!("Checking doc attributes in {}...", path);
-    for attr in krate.attrs {
-        if let syn::AttrStyle::Outer = attr.style {
-            continue;
-        }
-        let (ident, nested_meta_items) = match attr.interpret_meta() {
-            Some(syn::Meta::List(syn::MetaList { ident, nested, .. })) => (ident, nested),
-            _ => continue,
-        };
-
-        if ident != "doc" {
-            continue;
-        }
-
-        for nested_meta_item in nested_meta_items {
-            let meta_item = match nested_meta_item {
-                syn::NestedMeta::Meta(ref meta_item) => meta_item,
-                _ => continue,
-            };
-
-            let check_result = match *meta_item {
-                syn::Meta::NameValue(syn::MetaNameValue {
-                    ref ident, ref lit, ..
-                })
-                    if ident == "html_root_url" =>
-                {
-                    match *lit {
-                        // Accept both cooked and raw strings here.
-                        syn::Lit::Str(ref s) => url_matches(&s.value(), pkg_name, &version),
-                        // A non-string html_root_url is probably an
-                        // error, but we leave this check to the
-                        // compiler.
-                        _ => continue,
-                    }
-                }
-                syn::Meta::Word(ref name) if name == "html_root_url" => {
-                    Err(String::from("html_root_url attribute without URL"))
-                }
-                _ => continue,
-            };
-
-            // FIXME: use line number from the syn crate when it
-            // preserves span information. Here we simply find the
-            // first source line that contains "html_root_url".
-            //
-            // We know such a line must exist since we would have
-            // continue the loop above if it wasn't present.
-            let (line_no, source_line) = code
-                .lines()
-                .enumerate()
-                .find(|&(_, line)| line.contains("html_root_url"))
-                .expect("html_root_url attribute not present");
-
-            match check_result {
-                Ok(()) => {
-                    println!("{} (line {}) ... ok", path, line_no + 1);
-                    return Ok(());
-                }
-                Err(err) => {
-                    println!("{} (line {}) ... {} in", path, line_no + 1, err);
-                    println!("{}\n", indent(source_line));
-                    return Err(format!("html_root_url errors in {}", path));
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
+mod html_root_url;
+pub use html_root_url::check_html_root_url;
 
 /// Assert that the `html_root_url` attribute is up to date.
 ///
@@ -451,6 +312,9 @@ macro_rules! assert_contains_regex {
 
 #[cfg(test)]
 mod tests {
+    use semver_parser::range::parse as parse_request;
+    use semver_parser::version::parse as parse_version;
+
     use super::*;
 
     mod test_version_matches_request {
@@ -545,167 +409,4 @@ mod tests {
             );
         }
     }
-
-    mod test_url_matches {
-        use super::*;
-
-        #[test]
-        fn good_url() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("https://docs.rs/foo/1.2.3", "foo", &ver),
-                Ok(())
-            );
-        }
-
-        #[test]
-        fn trailing_slash() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("https://docs.rs/foo/1.2.3/", "foo", &ver),
-                Ok(())
-            );
-        }
-
-        #[test]
-        fn without_patch() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(url_matches("https://docs.rs/foo/1.2", "foo", &ver), Ok(()));
-        }
-
-        #[test]
-        fn without_minor() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(url_matches("https://docs.rs/foo/1", "foo", &ver), Ok(()));
-        }
-
-        #[test]
-        fn different_domain() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(url_matches("https://example.net/foo/", "bar", &ver), Ok(()));
-        }
-
-        #[test]
-        fn different_domain_http() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("http://example.net/foo/1.2.3", "foo", &ver),
-                Ok(())
-            );
-        }
-
-        #[test]
-        fn http_url() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("http://docs.rs/foo/1.2.3", "foo", &ver),
-                Err(String::from("expected \"https\", found \"http\""))
-            );
-        }
-
-        #[test]
-        fn bad_scheme() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("mailto:foo@example.net", "foo", &ver),
-                Err(String::from("expected \"https\", found \"mailto\""))
-            );
-        }
-
-        #[test]
-        fn no_package() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("https://docs.rs", "foo", &ver),
-                Err(String::from("missing package name"))
-            );
-        }
-
-        #[test]
-        fn no_package_trailing_slash() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("https://docs.rs/", "foo", &ver),
-                Err(String::from("missing package name"))
-            );
-        }
-
-        #[test]
-        fn no_version() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("https://docs.rs/foo", "foo", &ver),
-                Err(String::from("missing version number"))
-            );
-        }
-
-        #[test]
-        fn no_version_trailing_slash() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("https://docs.rs/foo/", "foo", &ver),
-                Err(String::from("missing version number"))
-            );
-        }
-
-        #[test]
-        fn bad_url() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("docs.rs/foo/bar", "foo", &ver),
-                Err(String::from("parse error: relative URL without a base"))
-            );
-        }
-
-        #[test]
-        fn bad_pkg_version() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("https://docs.rs/foo/1.2.bad/", "foo", &ver),
-                Err(String::from(
-                    "could not parse version in URL: \
-                     encountered unexpected token: AlphaNumeric(\"bad\")"
-                ))
-            );
-        }
-
-        #[test]
-        fn wrong_pkg_name() {
-            let ver = parse_version("1.2.3").unwrap();
-            assert_eq!(
-                url_matches("https://docs.rs/foo/1.2.3/", "bar", &ver),
-                Err(String::from("expected package \"bar\", found \"foo\""))
-            );
-        }
-    }
-
-    mod test_check_html_root_url {
-        use super::*;
-
-        #[test]
-        fn bad_path() {
-            let no_such_file = if cfg!(unix) {
-                "No such file or directory (os error 2)"
-            } else {
-                "The system cannot find the file specified. (os error 2)"
-            };
-            let errmsg = format!("could not read no-such-file.md: {}", no_such_file);
-            assert_eq!(
-                check_html_root_url("no-such-file.md", "foobar", "1.2.3"),
-                Err(errmsg)
-            );
-        }
-
-        #[test]
-        fn bad_pkg_version() {
-            // This uses the src/lib.rs file from this crate.
-            assert_eq!(
-                check_html_root_url("src/lib.rs", "foobar", "1.2"),
-                Err(String::from(
-                    "bad package version \"1.2\": expected more input"
-                ))
-            );
-        }
-    }
-
 }
